+++ src/lib.rs
@@ -25,7 +25,6 @@ use crate::events::{Event, EVENT_QUEUE};
 
 use skyline::libc::mkdir;
 use skyline::nro::{self, NroInfo};
-use skyline::{hooks::InlineCtx};
 use std::fs;
 
 use crate::menu::quick_menu_loop;
@@ -53,266 +52,6 @@ macro_rules! c_str {
         [$l.as_bytes(), "\u{0}".as_bytes()].concat().as_ptr()
     };
 }
-
-#[derive(Debug)]
-pub struct TValue {
-    value: u64,
-    tt: i32
-}
-
-#[skyline::hook(offset = 0x38f3d60)]
-pub unsafe fn handle_lua_setfield(
-    lua_state: u64,
-    lua_tvalue: *const TValue,
-    field_name: *const skyline::libc::c_char
-) {
-    if skyline::from_c_str(field_name) == "LayoutRootList" {
-        println!("In LayoutRootList");
-    }
-    original!()(lua_state, lua_tvalue, field_name);
-}
-
-#[skyline::hook(offset = 0x3777130)]
-pub unsafe fn handle_play_animation(
-    layout_view: u64,
-    animation_name: *const skyline::libc::c_char
-) -> u64 {
-    println!("play_animation: {}", skyline::from_c_str(animation_name));
-    original!()(layout_view, animation_name)
-}
-
-#[skyline::hook(offset = 0x3776cd0)]
-pub unsafe fn handle_play_animation_at_speed(
-    speed: f32,
-    unk: u64,
-    animation_name: *const skyline::libc::c_char
-) -> u64 {
-    println!("play_animation_at_speed: {}", skyline::from_c_str(animation_name));
-    original!()(speed, unk, animation_name)
-}
-
-#[skyline::hook(offset = 0x3777000)]
-pub unsafe fn handle_play_animation_at_speed2(
-    speed: f32,
-    unk: u64,
-    animation_name: *const skyline::libc::c_char
-) -> u64 {
-    println!("play_animation_at_speed2: {}", skyline::from_c_str(animation_name));
-    original!()(speed, unk, animation_name)
-}
-
-
-
-#[skyline::hook(offset = 0x3776ab0, inline)]
-pub unsafe fn handle_get_pane_animation(ctx: &mut InlineCtx) {
-    println!("get_pane_animation: {}", skyline::from_c_str(*ctx.registers[1].x.as_ref() as *const u8));
-}
-
-
-#[skyline::hook(offset = 0x4b120)]
-pub unsafe fn handle_bind_animation(
-    layout_view: u64,
-    animation_name: *const skyline::libc::c_char
-) -> u64 {
-    println!("bind_animation: {}", skyline::from_c_str(animation_name));
-    original!()(layout_view, animation_name)
-}
-
-#[skyline::hook(offset = 0x0595d0)]
-pub unsafe fn handle_bind_animation2(
-    layout_view: u64,
-    animation_name: *const skyline::libc::c_char,
-    unk1: u32,
-    unk2: u32
-) -> u64 {
-    println!("bind_animation: {}", skyline::from_c_str(animation_name));
-    original!()(layout_view, animation_name, unk1, unk2)
-}
-
-#[repr(C)]
-#[derive(Debug)]
-pub struct LayoutPaneUi2d {
-    unk_addresses: [u64; 6],
-    pos_x: f32,
-    pos_y: f32,
-    pos_z: f32,
-    rot_x: f32,
-    rot_y: f32,
-    rot_z: f32,
-    scale_x: f32,
-    scale_y: f32,
-    size_x: f32,
-    size_y: f32,
-    flags: u8,
-    alpha: u8
-}
-
-#[repr(C)]
-
-#[derive(Debug)]
-
-pub struct LayoutPane {
-    layout_pane_ui2d: *mut LayoutPaneUi2d,
-    picture: u64,
-    sub_layout_pane_user_data_unk: u64,
-    sub_layout_pane: *mut LayoutPane,
-}
-
-#[skyline::hook(offset = 0x3775480, inline)]
-pub unsafe fn handle_get_pane_by_name(
-    ctx: &mut InlineCtx
-) {
-    // Grabbing stuff off the stack is interesting. 
-    let pane_name = skyline::from_c_str(
-        (ctx as *const InlineCtx as *const u8).add(0x100).add(0xD8)
-    );
-    println!("get_pane_by_name: {}", pane_name); 
-    if pane_name == "set_dmg_p" || true {
-        let layout_pane = (*ctx.registers[0].x.as_ref()) as *mut LayoutPane;
-        if !layout_pane.is_null() {
-            println!("pane: {:#?}", *layout_pane);
-            // pane_set_text_string(layout_pane, c_str!("Test!"));
-            let sublayout_pane = (*layout_pane).sub_layout_pane;
-            if !sublayout_pane.is_null() {
-                println!("sublayout_pane: {:#?}", *sublayout_pane);
-                // pane_set_text_string(layout_pane, c_str!("Test!"));
-            }
-            let layout_pane_ui2d = (*layout_pane).layout_pane_ui2d;
-            if !layout_pane_ui2d.is_null() {
-                println!("pane_ui2d: {:#?}", *layout_pane_ui2d);
-                // Turn invisible
-                (*layout_pane_ui2d).scale_x = (*layout_pane_ui2d).scale_x * 2.0;
-                (*layout_pane_ui2d).scale_y = (*layout_pane_ui2d).scale_y * 2.0;
-                (*layout_pane_ui2d).flags = (*layout_pane_ui2d).flags | 0x10;
-
-            }
-        }
-    }
-}
-
-
-#[skyline::hook(offset = 0x3774ac0)]
-pub unsafe fn handle_set_enable_input(
-    layout_root: u64,
-    enable: bool
-) -> u64 {
-    println!("set_enable_input");
-    original!()(layout_root, enable)
-}
-
-pub struct AnimTransform {
-    vtable: u64,
-    unk: [u64; 2],
-    enabled: bool
-}
-
-pub struct AnimTransformNode {
-    data: *mut AnimTransform,
-    next: *mut AnimTransformNode
-}
-
-pub struct RawLayout {
-    anim_trans_list: AnimTransformNode,
-    root_pane: *const LayoutPaneUi2d,
-    group_container: u64,
-    layout_size: f64,
-    layout_name: *const skyline::libc::c_char
-}
-
-pub struct Layout {
-    vtable: u64,
-    raw_layout: RawLayout
-}
-
-#[skyline::hook(offset = 0x4b620)]
-pub unsafe fn handle_draw(layout: *mut Layout, draw_info: u64, cmd_buffer: u64) {
-    let layout_name = skyline::from_c_str((*layout).raw_layout.layout_name);
-    let layout_root_pane = (*layout).raw_layout.root_pane;
-    let mut curr = &mut (*layout).raw_layout.anim_trans_list as *mut AnimTransformNode;
-
-    if layout_name == "info_training" {
-        for s in vec![
-            "txt_cap_00",
-            "set_txt_num_00",
-            "set_txt_num_01",
-        ] {
-            let txt_pane = find_pane_by_name_recursive(layout_root_pane, c_str!(s));
-            // println!("Replacing {}/{}...", layout_name, s);
-            pane_set_text_string(txt_pane, c_str!("Hello!"));            
-            // println!("Txt Pane: {:#X?}", *txt_pane);
-        }
-    }
-
-    if layout_name == "info_melee" {
-        for s in vec![
-            "p1"
-        ] {
-            let dmg_pane = find_pane_by_name_recursive(layout_root_pane, c_str!(s));
-            (*dmg_pane).pos_y = (*dmg_pane).pos_y + 300.0;
-            for anim_search_name in vec![
-                // "dig_3",
-                // "dig_3_anim",
-                // "dig_3_reach",
-                // "set_dmg_num_3",
-                // "dig_2",
-                // "dig_2_anim",
-                // "dig_2_reach",
-                // "set_dmg_num_2",
-                // "dig_1",
-                // "dig_1_anim",
-                // "dig_1_reach",
-                // "set_dmg_num_1",
-                // "dig_0",
-                // "dig_0_anim",
-                // "dig_0_reach",
-                // "set_dmg_num_0",
-                // "set_dmg_num_dec",
-                // "dig_dec",
-                // "dig_dec_reach_0",
-                // "dig_dec_anim_00",
-                // "dig_dec_reach_1",
-                // "dig_dec_reach_01",
-                "set_dmg_p",
-                "changedig"
-            ] {
-                let dmg_pane_p1 = find_pane_by_name_recursive(dmg_pane, c_str!(anim_search_name));
-                if !dmg_pane_p1.is_null() {
-                    println!("Found pane by {}::find_pane_by_name({}): {:X?}", layout_name, anim_search_name, *dmg_pane_p1);
-                }
-            }
-        }
-    }
-
-
-    
-    original!()(layout, draw_info, cmd_buffer);
-}
-
-#[skyline::hook(offset = 0x4b120)] 
-pub unsafe fn handle_pane_bind_animation(layout: *mut Layout, anim: *const u64) {
-    println!("Bind Animation");
-    original!()(layout, anim)
-}
-
-#[skyline::from_offset(0x59970)]
-pub unsafe fn find_pane_by_name_recursive(
-    pane: *const LayoutPaneUi2d,
-    s: *const skyline::libc::c_char
-) -> *mut LayoutPaneUi2d;
-
-#[skyline::from_offset(0x583c0)]
-pub unsafe fn find_pane_by_name(
-    pane: *const LayoutPaneUi2d,
-    s: *const skyline::libc::c_char,
-    some_bool_maybe: bool
-) -> *mut LayoutPaneUi2d;
-
-#[skyline::from_offset(0x37a1270)]
-pub unsafe fn pane_set_text_string(
-    pane: *const LayoutPaneUi2d,
-    s: *const skyline::libc::c_char
-);
-
 
 #[skyline::main(name = "training_modpack")]
 pub fn main() {
@@ -346,19 +85,7 @@ pub fn main() {
         EVENT_QUEUE.push(Event::smash_open());
     }
 
-    skyline::install_hooks!(
-        // handle_lua_setfield,
-        // handle_play_animation,
-        // handle_play_animation_at_speed,
-        // handle_get_pane_animation,
-        // handle_play_animation_at_speed2,
-        // handle_bind_animation,
-        // handle_bind_animation2,
-        // handle_set_enable_input,
-        // handle_get_pane_by_name,
-        handle_draw,
-        handle_pane_bind_animation
-    );
+    training::ui::install_hooks();
 
     hitbox_visualizer::hitbox_visualization();
     hazard_manager::hazard_manager();
