+++ src/training/mod.rs
@@ -23,6 +23,7 @@ pub mod sdi;
 pub mod shield;
 pub mod tech;
 pub mod throw;
+pub mod ui;
 
 mod air_dodge_direction;
 mod attack_angle;
